Passing arguments simplified in nested components
