+++ src/components/SaleComponent/SaleTabs.js
@@ -1,27 +1,39 @@
 import React, { useState } from "react";
 import '../../styles/Tabs.css';
+import ProductService from "../../services/ProductService";
+import SalesService from "../../services/SalesService";
 import ServiceSaleTab from "./ServiceSaleTab";
 import ProductSaleTab from "./ProductSaleTab";
 
-const Tabs = ({ 
-  services,
-  products,
-  staff,
-  staffMember,
-  service,
-  product,
-  date,
-  addSale,
-  updateStock,
-  handleServiceChange,
-  handleProductChange,
-  handleStaffMemberChange,
-  handleDateChange}) => {
+const SaleTabs = ({ services, products, staff, sales, setSales }) => {
 
   const [activeTab, setActiveTab] = useState("tab1");
+  const [service, setService] = useState();
+  const [product, setProduct] = useState();
+  const [staffMember, setStaffMember] = useState();
+  const [date, setDate] = useState();    
 
   function handleTab1() {setActiveTab("tab1")};
   function handleTab2() {setActiveTab("tab2")};
+
+  const handleServiceChange = (e) => {setService(services[e.target.value])}
+  const handleProductChange = (e) => {setProduct(products[e.target.value])}
+  const handleStaffMemberChange = (e) => {setStaffMember(staff[e.target.value])}
+  const handleDateChange = (e) => {setDate(e.target.value)}
+
+  const updateStock = (stockSold) => {
+    const updatedProduct = {...product}
+    updatedProduct.stock -= 1
+    updatedProduct.sold += 1
+    setProduct(updatedProduct)
+    ProductService.updateProduct(updatedProduct)
+  }
+
+  const addSale = (sale) => {
+    SalesService.addSale(sale);
+    const updatedSales = [...sales, sale]  
+    setSales(updatedSales);
+  }
 
   return (
     <div className="Tabs child-container">
@@ -57,4 +69,4 @@ const Tabs = ({
   );
 };
 
-export default Tabs;
+export default SaleTabs;
