+++ src/containers/Sales.js
@@ -9,12 +9,8 @@ import ProductService from '../services/ProductService';
 const Sales = () => {
     const [sales, setSales] = useState([]);
     const [services, setServices] = useState([]);
-    const [service, setService] = useState();
     const [products, setProducts] = useState([]);
-    const [product, setProduct] = useState();
     const [staff, setStaff] = useState([]);
-    const [staffMember, setStaffMember] = useState();
-    const [date, setDate] = useState();    
 
     useEffect(() => {
         SalesService.getSales()
@@ -33,41 +29,14 @@ const Sales = () => {
             .then(staff => setStaff(staff));
     }, []);
 
-    const handleServiceChange = (e) => {setService(services[e.target.value])}
-    const handleProductChange = (e) => {setProduct(products[e.target.value])}
-    const handleStaffMemberChange = (e) => {setStaffMember(staff[e.target.value])}
-    const handleDateChange = (e) => {setDate(e.target.value)}
-
-    const updateStock = (stockSold) => {
-        const updatedProduct = {...product}
-        updatedProduct.stock -= 1
-        updatedProduct.sold += 1
-        setProduct(updatedProduct)
-        ProductService.updateProduct(updatedProduct)
-    }
-    
-    const addSale = (sale) => {
-        SalesService.addSale(sale);
-        const updatedSales = [...sales, sale]  
-        setSales(updatedSales);
-    }
-
     return (
         <div className="parent-container">
             <SaleTabs 
                 services={services}
                 products={products}
                 staff={staff}
-                staffMember={staffMember}
-                service={service}
-                product={product}
-                date={date}
-                addSale={addSale}
-                updateStock={updateStock}
-                handleServiceChange={handleServiceChange}
-                handleProductChange={handleProductChange}
-                handleStaffMemberChange={handleStaffMemberChange}
-                handleDateChange={handleDateChange} />
+                sales={sales}
+                setSales={setSales} />
             <SalesTable data={sales}/>
         </div>
     )
